webinar-materials/starter-kits/customer-support-chatbot/requirements.txt
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
python-dotenv==1.1.1
pandas==2.3.2
chromadb==1.0.20
